third_party/WebKit/LayoutTests/fast/canvas/imagedata-contains-uint8clampedarray.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function(){

    var canvas = document.createElement("canvas");
    canvas.width = 10;
    canvas.height = 10;
    var context = canvas.getContext("2d");

    var imageData = context.getImageData(0, 0, 1, 1);
    assert_equals(imageData.data.constructor, Uint8ClampedArray);

}, "Tests that ImageData contains a Uint8ClampedArray.");
</script>
